Bring Adw.Dialog callback logic to one place
